[ARM PATCH] 2563/1: RiscPC - update IOMEM annotations
Patch from Ben Dooks

Fixes `iomem` annotations on a number of parts
of the arch-rpc build, including IO_BASE, IOC_BASE
and IOMD_BASE.
Changed the include/asm-arm/arch-rpc/io.h and hardware.h
to reflect this, and fixed-up the build problems

+++ arch/arm/kernel/ecard.c
@@ -228,7 +228,7 @@ static void ecard_init_pgtables(struct mm_struct *mm)
 	 */
 	pgd_t *src_pgd, *dst_pgd;
 
-	src_pgd = pgd_offset(mm, IO_BASE);
+	src_pgd = pgd_offset(mm, (unsigned long)IO_BASE);
 	dst_pgd = pgd_offset(mm, IO_START);
 
 	memcpy(dst_pgd, src_pgd, sizeof(pgd_t) * (IO_SIZE / PGDIR_SIZE));
